requirements.txt
paddlepaddle
paddleocr
matplotlib
opencv-python
spacy
mysql-connector-python
pillow